Now a working game with Slick.
The game does now works. It spawns a spaceship in the middle of the game
and spawns ten asteroids at random locations. If you fly out at one side
of the game, you will appear at the other side. However, there is no
collision detection yet or possibility to fire bullets. And yeah, it's
possible to control the spaceship with the keyboard. I have also removed
the DPoint class since there is already one that works like that in the
Java API.

Change-Id: I6db9e7a1875ace8772ab8f21183f1a6ebced0b2a
Signed-off-by: Daniel "MaTachi" Jonsson <m@danielj.se>